requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
botasaurus==3.2.3
botasaurus-proxy-authentication==1.0.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
chromedriver-autoinstaller-fix==1.0.4
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cryptography==41.0.7
dataclasses-json==0.6.3
defusedxml==0.7.1
distro==1.8.0
dnspython==2.4.2
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
javascript==1!1.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kaitaistruct==0.10
langchain==0.0.352
langchain-anthropic==0.1.4
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
linkedin-scraper==2.11.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.6.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
proxycurl==0.0.13
proxycurl-py==0.1.0.post1
psutil==5.9.7
pyarrow==14.0.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydivert==2.1.0
Pygments==2.17.2
pymongo==4.6.1
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.15.2
selenium==4.5.0
selenium-wire==5.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
streamlit==1.30.0
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
vulture==2.10
watchdog==3.0.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
